Fix CI pipeline failures: resolve audit vulnerabilities, PHPStan type errors, and test regressions

+++ app/Actions/Workouts/FetchWorkoutsIndexAction.php
@@ -93,7 +93,7 @@ class FetchWorkoutsIndexAction
                     instanceof Carbon
                     ? $date->format('M')
                     : '',
-                'count' => count($rows),
+                'count' => (int) count($rows),
             ])->values();
     }
